+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- No longer fails when a service is not provided with a directory endpoint.
+
 ## v2.17.3 - 2023-08-21
 
 ### Fixed